+++ .vs/VSWorkspaceState.json
@@ -1,8 +1,8 @@
 {
   "ExpandedNodes": [
     "",
-    "\\TestProject1"
+    "\\TestProject2"
   ],
-  "SelectedNode": "\\TestProject1\\UnitTest1.cs",
+  "SelectedNode": "\\dekoder.cs",
   "PreviewInSolutionExplorer": false
 }
