+++ main.cs
@@ -1,44 +1,114 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+
 
 namespace morseovka
 {
-    class Program
+    public class Morseovka
     {
+
         static void Main(string[] args)
         {
-            Console.WriteLine("Morse code");
-            //smyčka pro případné opakovaní funkce
-            int ret = 1;
-            while (ret == 1)
-            {
-                Console.Write("Press 1 if you want to encode ot press 2 if you want to decode:  ");
-            again:
-                //zde se program zeptá, co chcete aby dělal
-                int num = Convert.ToInt32(Console.ReadLine());
-                //jesti že chcete spustit encode spustí se tenhle if:
-                if (num == 1)
-                {
-                    //spusteni encode 
-                    encoder mors = new encoder();
-                    mors.StartEncode();
-                }
-                //nebo jesti chcete spustit decode spustí se tenhle elseif:
-                else if (num == 2)
-                {
-                    //spusteni decode programu
-                    decoder mors = new decoder();
-                    mors.StartDecode();
-                }
-                //jinak vypíše chybné zapsání čísla a pošle uživatele zpátky na začátek programu
-                else
-                {
-                    Console.WriteLine("you entered wrong number:{0}", num);
-                    goto again;
-                }
-                //jestli že chce uživatel znovu použít program napíše 1
-                Console.Write("press 1 if you want another en/decode, otherwise press 2:  ");
-                ret = Convert.ToInt32(Console.ReadLine());
-            }
+            Console.WriteLine("write morse code you want to encode:  ");
+            string prelozm = Console.ReadLine();
+            prelozm = prelozm.ToLower();
+            Console.WriteLine(zakoduj(prelozm));
+
+
+            Console.WriteLine("write morse code you want to decode:  ");
+            string preloz = Console.ReadLine();
+            Console.WriteLine(Dekoduj(preloz));
         }
+        public static string zakoduj(string zkod)
+        {
+
+            var text = zkod;
+
+            var m = string.Empty;
+            foreach (var c in text)
+            {
+                m += morseovka[c] + " ";
+            }
+            return m;
+
+        }
+
+        public static string Dekoduj(string mkod)
+        {
+            string[] mvety = mkod.Split("    ");
+            string prelozeno = "";
+
+            foreach (string mslovo in mvety)
+            {
+                string[] mpismena = mslovo.Split(" ");
+                string prelozenoslovo = "";
+
+                foreach (string mpismeno in mpismena)
+                {
+                    prelozenoslovo += morseovka.FirstOrDefault(x => x.Value == mpismeno).Key;
+                    
+                }
+
+                prelozeno += (prelozeno == "") ? prelozenoslovo : " " + prelozenoslovo;
+
+            }
+            return prelozeno;
+        }
+        public static Dictionary<char, string> morseovka = new Dictionary<char, string>()
+        {
+
+             {' ' , "/"},
+             {'a' , ".-"},
+             {'b' , "-..."},
+             {'c' , "-.-."},
+             {'d' , "-.."},
+             {'e' , "."},
+             {'f' , "..-."},
+             {'g' , "--."},
+             {'h' , "...."},
+             {'i' , ".."},
+             {'j' , ".---"},
+             {'k' , "-.-"},
+             {'l' , ".-.."},
+             {'m' , "--"},
+             {'n' , "-."},
+             {'o' , "---"},
+             {'p' , ".--."},
+             {'q' , "--.-"},
+             {'r' , ".-."},
+             {'s' , "..."},
+             {'t' , "-"},
+             {'u' , "..-"},
+             {'v' , "...-"},
+             {'w' , ".--"},
+             {'x' , "-..-"},
+             {'y' , "-.--"},
+             {'z' , "--.."},
+             {'0' , "-----"},
+             {'1' , ".----"},
+             {'2' , "..---"},
+             {'3' , "...--"},
+             {'4' , "....-"},
+             {'5' , "....."},
+             {'6' , "-...."},
+             {'7' , "--..."},
+             {'8' , "---.."},
+             {'9' , "----."},
+             {'?' , "..--.."},
+             {';' , "-.-.-."},
+             {'.' , ".-.-.-"},
+             {',' , "--..--"},
+             {'/' , "-..-."},
+             {'=' , "-...-"},
+             {'-' , "-....-"},
+             {'(' , "-.--."},
+             {')' , "-.--.-"},
+             {'"' , ".-..-."},
+             {':' , "---..."},
+             {'_' , "..--.-"},
+             {'+' , ".-.-."},
+             {'@' , ".--.-."},
+        };
     }
 }
